Improvements to getting scores of games.
Improved the approach for when to look up scores.

Fixed resource leak (cleaned up after scanning the web page).

Made improvement for user selection of games that have started.
Since we are using the scheduling page for game updates, and that
page rarely updates during a game, the "game.Status" will say Future.
So logic included that if selection is made after game start time,
the selection is ignored.